+++ src/main/java/eu/fayder/restcountries/domain/ICountryRestSymbols.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package eu.fayder.restcountries.v3.domain;
+package eu.fayder.restcountries.domain;
 
 public interface ICountryRestSymbols {
     public static final String SEMICOLON = ";";
