+++ src/main/kotlin/eu/fayder/restcountries/v3/domain/Country.kt
@@ -1,5 +1,7 @@
 package eu.fayder.restcountries.v3.domain
 
+import eu.fayder.restcountries.domain.BaseCountry
+
 data class Country
 (
         val currencies: List<Currency>?,
